+++ lib/overlay.js
@@ -4,6 +4,7 @@ import styled from 'styled-components';
 import { Button } from './button';
 import { Title, Info, Actions } from './block';
 import { useCallbackProxy } from './callback-proxy';
+import { useEscape, useFocusTrap } from './keyboard-navigation';
 
 const OverlayWrap = styled.div`
   position: fixed;
@@ -65,46 +66,12 @@ const FixedBody = () => {
   return scrollY === null ? null : <FixedBodyStyle scrollY={scrollY} />;
 };
 
-const useFocusTrap = () => {
-  const first = React.useRef();
-  const last = React.useRef();
-  React.useEffect(() => {
-    const handler = (e) => {
-      if (e.key === 'Tab' && !e.shiftKey && e.target === last.current) {
-        e.preventDefault();
-        first.current.focus();
-      }
-      if (e.key === 'Tab' && e.shiftKey && e.target === first.current) {
-        e.preventDefault();
-        last.current.focus();
-      }
-    };
-    window.addEventListener('keydown', handler);
-    return () => window.removeEventListener('keydown', handler);
-  }, []);
-  return { first, last };
-};
-
 const useFocusOnMount = () => {
   const ref = React.useRef();
   React.useEffect(() => {
     ref.current.focus();
   }, []);
   return ref;
-};
-
-const useEscape = (onClose) => {
-  const memoOnClose = useCallbackProxy(onClose);
-
-  React.useEffect(() => {
-    const handler = (e) => {
-      if (e.key === 'Escape') {
-        memoOnClose(e);
-      }
-    };
-    window.addEventListener('keydown', handler);
-    return () => window.removeEventListener('keydown', handler);
-  }, [memoOnClose]);
 };
 
 // https://www.w3.org/TR/wai-aria-practices-1.1/#dialog_modal
