+++ lib/popover.js
@@ -4,6 +4,7 @@ import styled, { css } from 'styled-components';
 import { Button } from './button';
 import { Title, Info, Actions, DangerZone } from './block';
 import { useCallbackProxy } from './callback-proxy';
+import { useEscape } from './keyboard-navigation';
 
 const debounce = (fn, timeout) => {
   let handle;
@@ -145,17 +146,22 @@ const useStack = (defaultState) => {
 
 export const PopoverMenu = ({ label, buttonProps = {}, views = {}, initialView, children, ...props }) => {
   const [open, setOpen] = React.useState(false);
+  const focusOnOpenRef = React.useRef();
   const { top: activeView, push: setActiveView, pop: onBack, clear } = useStack(initialView);
   const activeViewFunc = views[activeView] || children;
   const onOpen = () => {
     setOpen(true);
+    if (focusOnOpenRef.current) focusOnOpenRef.current.focus();
     clear();
   };
   const onClose = () => {
     setOpen(false);
     clear();
   };
-
+  React.useEffect(() => {
+    if (focusOnOpenRef.current) focusOnOpenRef.current.focus();
+  }, [activeView])
+  
   const rootRef = useClickOutside(open, onClose);
 
   const onToggle = (e) => {
@@ -166,17 +172,16 @@ export const PopoverMenu = ({ label, buttonProps = {}, views = {}, initialView, 
       onClose();
     }
   };
-  const onKeyDown = (e) => {
-    if (e.key === 'Escape') onClose();
-  };
+  
+  useEscape(onClose);
 
   // see https://www.w3.org/TR/wai-aria-practices-1.1/#menubutton
   return (
     <PopoverContainer data-module="PopoverMenu" ref={rootRef} onToggle={onToggle} open={open} role="menu" {...props}>
-      <PopoverToggle onKeyDown={onKeyDown} aria-haspopup="menu" aria-expanded={open}>
+      <PopoverToggle aria-haspopup="menu" aria-expanded={open}>
         {label}
       </PopoverToggle>
-      {activeViewFunc({ open, activeView, setActiveView, onClose, onBack })}
+      {activeViewFunc({ open, activeView, setActiveView, onClose, onBack, focusOnOpenRef })}
     </PopoverContainer>
   );
 };
@@ -216,32 +221,32 @@ export const StoryPopoverMenu_Multiple_Views = () => (
   <PopoverMenu
     label={<Button as="span">Open Multi-Page Popover</Button>}
     views={{
-      foo: ({ onClose, onBack }) => (
+      foo: ({ onClose, onBack, focusOnOpenRef }) => (
         <Popover align="left" size="wide">
-          <Title onBack={onBack} onClose={onClose}>
+          <Title onBack={onBack} onBackRef={focusOnOpenRef} onClose={onClose}>
             Foo
           </Title>
           <Info>Foo content goes here</Info>
         </Popover>
       ),
-      bar: ({ setActiveView, onClose, onBack }) => (
+      bar: ({ setActiveView, onClose, onBack, focusOnOpenRef }) => (
         <Popover align="left" size="wide">
-          <Title onBack={onBack} onClose={onClose}>
+          <Title onBack={onBack}  onClose={onClose}>
             Bar
           </Title>
           <Info>Bar content goes here</Info>
           <Actions>
-            <Button onClick={() => setActiveView('foo')}>Show Foo</Button>
+            <Button ref={focusOnOpenRef} onClick={() => setActiveView('foo')}>Show Foo</Button>
           </Actions>
         </Popover>
       ),
     }}
   >
-    {({ setActiveView, onClose }) => (
+    {({ setActiveView, onClose, focusOnOpenRef }) => (
       <Popover align="left" size="wide">
         <Title onClose={onClose}>Multi-Page Popover</Title>
         <ActionsStack>
-          <Button onClick={() => setActiveView('foo')}>Show Foo</Button>
+          <Button ref={focusOnOpenRef} onClick={() => setActiveView('foo')}>Show Foo</Button>
           <Button onClick={() => setActiveView('bar')}>Show Bar</Button>
         </ActionsStack>
       </Popover>
